More testing (#48)
* Added some casting for py3.11 and py3.12 since np.count returns int64.
* Enabled some xfail tests, fixed bugs in tests or in code, and filed issues for seemingly real bugs.
* uint32/64 issue with np.count needed to be propagated to all census-ing components.
* Use verified baseline model as "ref model" in comparison tests.

+++ src/laser_generic/transmission.py
@@ -379,7 +379,7 @@ class TransmissionSIR(Transmission):
         condition = population.susceptibility[0 : population.count] == 0
 
         if len(patches) == 1:
-            np.add(contagion, np.sum(condition), out=contagion)
+            np.add(contagion, np.uint32(np.sum(condition)), out=contagion)
         else:
             nodeids = population.nodeid[0 : population.count]
             np.add.at(contagion, nodeids[condition], 1)
